configurando serviço de descoberta com eureka

+++ meme_service/src/main/java/br/com/memelandia/service/ServiceMeme.java
@@ -14,6 +14,7 @@ import br.com.memelandia.entities.Usuario;
 import br.com.memelandia.repositori.RepositoriMeme;
 import br.com.memelandia.repositori.RepositoriCategoria;
 import br.com.memelandia.repositori.RepositoriUsuario;
+import io.micrometer.core.instrument.*;
 
 /**
  * @author fsdney
@@ -23,6 +24,9 @@ import br.com.memelandia.repositori.RepositoriUsuario;
 public class ServiceMeme {
 
  private Logger logger = LoggerFactory.getLogger(ServiceMeme.class);
+ 
+ @Autowired
+ private MeterRegistry meterRegistry; // Registro de métricas
 
  @Autowired
  private RepositoriMeme repositoriMeme;
@@ -34,22 +38,49 @@ public class ServiceMeme {
  private RepositoriUsuario repositoriUsuario;
 
  public List<Meme> listarTodosMemes() {
-     logger.info("Listando todos os memes.");
-     return repositoriMeme.findAll();
+	 logger.info("Recebida requisição para listar todos os memes.");
+	 
+	// Contador de chamadas ao método
+     meterRegistry.counter("meme.listar.todas.chamadas").increment();
+     
+     // Medição de tempo de execução
+     long startTime = System.currentTimeMillis();
+     
+     List<Meme> memes = repositoriMeme.findAll();
+     
+     long endTime = System.currentTimeMillis();
+     meterRegistry.timer("meme.listar.todas.tempo").record(endTime - startTime, java.util.concurrent.TimeUnit.MILLISECONDS);
+	 
+	 if(memes.isEmpty()) {
+		 logger.warn("A lista está vazia.");
+         meterRegistry.counter("meme.listar.todas.vazio").increment();
+	 } else {
+		 logger.info("Total de memes encontrados: {}", memes.size());
+     	 meterRegistry.counter("meme.listar.todas.sucesso").increment(); // Métrica para sucesso
+         meterRegistry.gauge("meme.listar.todas.tamanho", memes, List::size); // Tamanho da lista retornada
+	 }
+	 
+     return memes;
  }
 
  public Meme criarMeme(Meme meme) {
-     logger.info("Criando novo meme: {}", meme);
+	 logger.info("Recebida requisição para listar todos os memes.");
+	 
+	 meterRegistry.counter("meme.criar.chamadas").increment();
+	 
+     long startTime = System.currentTimeMillis();
 
      // Valida categoria
      Optional<Categoria> categoria = repositoriCategoria.findById(meme.getCategoria().getId());
      if (categoria.isEmpty()) {
+    	 logger.warn("Categoria não encontrada.");
          throw new RuntimeException("Categoria não encontrada.");
      }
 
      // Valida usuário
      Optional<Usuario> usuario = repositoriUsuario.findById(meme.getUsuario().getId());
      if (usuario.isEmpty()) {
+    	 logger.warn("O usuario não encontrado.");
          throw new RuntimeException("Usuário não encontrado.");
      }
 
@@ -57,18 +88,61 @@ public class ServiceMeme {
      meme.setDataCadastro(new java.sql.Date(System.currentTimeMillis()));
      Meme salvo = repositoriMeme.save(meme);
      logger.info("Meme criado com sucesso: {}", salvo);
+     meterRegistry.counter("meme.criar.sucesso").increment(); // Métrica para sucesso
+		
+		long endTime = System.currentTimeMillis();
+		meterRegistry.timer("meme.criar.tempo").record(endTime - startTime, java.util.concurrent.TimeUnit.MILLISECONDS);
+     
      return salvo;
  }
 
  public Optional<Meme> buscarMemePorId(Long id) {
-     logger.info("Buscando meme com ID: {}", id);
-     return repositoriMeme.findById(id);
+	 logger.info("Recebida requisição para buscar meme com ID: {}", id);
+	 
+	// Contador de chamadas ao método
+ 	meterRegistry.counter("meme.buscarPorID.chamadas").increment();
+
+ 	// Medição de tempo de execução
+ 	long startTime = System.currentTimeMillis();
+ 	
+ 	 Optional<Meme> meme = repositoriMeme.findById(id);
+     long endTime = System.currentTimeMillis();
+		meterRegistry.timer("meme.buscarPorID.tempo").record(endTime - startTime, java.util.concurrent.TimeUnit.MILLISECONDS);
+		
+     if (meme.isPresent()) {
+         logger.info("meme encontrado: {}", meme.get());
+         meterRegistry.counter("meme.buscarPorID.sucesso").increment(); // Métrica para sucesso
+     } else {
+         logger.warn("meme com ID {} não encontrado.", id);
+         meterRegistry.counter("meme.buscarPorID.naoEncontrada").increment(); // Métrica para falha
+     }
+	 
+     return meme;
  }
 
  public void deletarMeme(Long id) {
-     logger.info("Deletando meme com ID: {}", id);
-     repositoriMeme.deleteById(id);
+	 logger.info("Recebida requisição para deletar meme com ID: {}", id);
+	 
+	 // Contador de chamadas ao método
+	meterRegistry.counter("meme.deletar.chamadas").increment();
+	// Medição de tempo de execução
+     long startTime = System.currentTimeMillis();
+     
+     Optional<Meme> memeExistente = repositoriMeme.findById(id);
+     if (memeExistente.isPresent()) {
+     	repositoriMeme.deleteById(id);
+     	logger.info("meme com ID {} deletada com sucesso.", id);
+         meterRegistry.counter("meme.deletar.sucesso").increment(); // Métrica para sucesso
+     } else {
+     	logger.warn("meme com ID {} não encontrada.", id);
+         meterRegistry.counter("meme.deletar.naoEncontrada").increment(); // Métrica para falha
+     }
+     
+     long endTime = System.currentTimeMillis();
+		meterRegistry.timer("meme.deletar.tempo").record(endTime - startTime, java.util.concurrent.TimeUnit.MILLISECONDS);
+		
  }
+     
 
  
  public Meme obterMemeDoDia() {
